Move exporter config types to linera-service
