+++ linera-service/src/exporter/runloops/validator_exporter.rs
@@ -13,11 +13,11 @@ use std::{
 use futures::{future::try_join_all, stream::FuturesOrdered};
 use linera_base::identifiers::BlobId;
 use linera_chain::types::ConfirmedBlockCertificate;
-use linera_client::config::DestinationId;
 use linera_core::node::{
     CrossChainMessageDelivery, NodeError, ValidatorNode, ValidatorNodeProvider,
 };
 use linera_rpc::grpc::{GrpcClient, GrpcNodeProvider};
+use linera_service::config::DestinationId;
 use linera_storage::Storage;
 use tokio::sync::mpsc::{Receiver, Sender};
 use tokio_stream::StreamExt;
