+++ linera-service/src/exporter/state.rs
@@ -14,11 +14,11 @@ use linera_base::{
     data_types::BlockHeight,
     identifiers::{BlobId, ChainId},
 };
-use linera_client::config::DestinationId;
 use linera_sdk::{
     ensure,
     views::{RootView, View},
 };
+use linera_service::config::DestinationId;
 use linera_views::{
     context::Context, log_view::LogView, map_view::MapView, register_view::RegisterView,
     set_view::SetView, views::ClonableView,
