+++ linera-service/src/lib.rs
@@ -8,6 +8,7 @@
 
 pub mod cli;
 pub mod cli_wrappers;
+pub mod config;
 pub mod node_service;
 pub mod project;
 #[cfg(with_metrics)]
